Create TakeAway Component for Restaurant Detail Page
